video_analyzer/requirements.txt
google-generativeai
gspread
google-auth-oauthlib
google-auth-httplib2
gspread-dataframe
tqdm
Flask